requirements.txt
jira
PyYAML
